# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Places API",
      "version": "1.0.0",
      "description": "API for managing places and user-specific places."
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
    "paths": {
      "/places": {
        "get": {
          "summary": "Get all places",
          "responses": {
            "200": {
              "description": "A list of places",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "places": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/Place"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/books": {
        "get": {
          "summary": "Get all places",
          "responses": {
            "200": {
              "description": "A list of places",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "places": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/Place"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/user-places": {
        "get": {
          "summary": "Get user-specific places",
          "responses": {
            "200": {
              "description": "A list of user-specific places",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "places": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/Place"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "put": {
          "summary": "Add a place to user-specific places",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "placeId": {
                      "type": "string",
                      "description": "ID of the place to add"
                    }
                  },
                  "required": ["placeId"]
                }
              }
            },
            "responses": {
              "200": {
                "description": "Updated user-specific places",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "userPlaces": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Place"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/user-places/{id}": {
        "delete": {
          "summary": "Remove a place from user-specific places",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID of the place to remove",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Updated user-specific places after deletion",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "userPlaces": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/Place"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Place": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "description": "Unique identifier for the place"
            },
            "name": {
              "type": "string",
              "description": "Name of the place"
            },
            "description": {
              "type": "string",
              "description": "Description of the place"
            },
            "location": {
              "type": "object",
              "properties": {
                "latitude": {
                  "type": "number",
                  "format": "float"
                },
                "longitude": {
                  "type": "number",
                  "format": "float"
                }
              }
            }
          },
          "required": ["id", "name", "location"]
        }
      }
    }
  }